deps: upgrade rainbow from 3.0.0 to 4.0.0 (#47)
The `Rainbow.extractModes` method is now depreceted. We should use `Rainbow.extractModes(from:)`
instead. This returns a `Rainbow.Entry` object that contains the nodes. We implment an internal
`Spinner.entryLength` that returns the length of the `Rainbow.Entry` object to enable us to us
to calulate the padding of messages and animations.